Friends (#6)
* Added friends table

* Fixed stale function

* Fixed PR comments

* Removed display name functions

* Added tests for friends

* Changed SSID to uint64

* Fixed uint64 and removed unnecessary functions discussed in PR

* Made sizes of data fields consistent

* Added logging for DB

+++ db/conversations_test.go
@@ -12,9 +12,9 @@ func ConversationsTest(t *testing.T) {
 func ConversationsTestID(t *testing.T) {
 	conversations := GetConversationUsers("karateAMD", "10.192.345.987")
 	assert.Equal(t, 3, len(conversations))
-	assert.Equal(t, 34, int(conversations[0].Session.SSID))
-	assert.Equal(t, 34, int(conversations[1].Message.SSID))
-	assert.Equal(t, 34, int(conversations[2].Session.SSID))
+	assert.Equal(t, uint64(34), conversations[0].Session.SSID)
+	assert.Equal(t, uint64(34), conversations[1].Message.SSID)
+	assert.Equal(t, uint64(34), conversations[2].Session.SSID)
 	assert.Equal(t, "2018-04-10 12:30:08.222222", conversations[0].Message.Timestamp)
 	assert.Equal(t, "When are we playing Fortnite?", conversations[2].Message.Text)
 }
